+++ __tests__/html/cardAction.adaptiveCard.cardAgenda.js
@@ -1,0 +1,7 @@
+/**
+ * @jest-environment ./__tests__/html/__jest__/WebChatEnvironment.js
+ */
+
+describe('"Adaptive card', () => {
+  test('card agenda should fail', () => runHTMLTest('cardAction.adaptiveCard.cardAgenda.html'));
+});