+++ packages/api/src/hooks/middleware/applyMiddleware.js
@@ -10,17 +10,21 @@ export default function applyMiddleware(type, ...middleware) {
       throw new Error(`reached terminator of ${type}`);
     });
 }
-
+/**
+ *
+ * @param {string} type Required. String equivalent of type of container to be rendered.
+ * @param { strict = false } - Used to enforce new middleware format which cooperates with new activity grouping.
+ * @see See {@link https://github.com/microsoft/BotFramework-WebChat/blob/master/CHANGELOG.md#4100---2020-08-18} and {@link https://github.com/microsoft/BotFramework-WebChat/pull/3365} for middleware breaking changes.
+ * @param  {middleware[]} middleware list of middleware to be applied.
+ * 'createRendererArgs' is "what to render"; for example, an activity.
+ * @returns  Returns a function if there is a renderer *committed* to render OR returns false if nothing should be rendered.
+ */
 export function forRenderer(type, { strict = false } = {}, ...middleware) {
   return (...setupArgs) => {
     const runMiddleware = concatMiddleware(...middleware)(...setupArgs)(() => (
       <ErrorBox error={new Error(`reached terminator of ${type}`)} type={type} />
     ));
 
-    // The createRendererArgs is "what to render", for example, activity.
-    // The function should return with only one of the two results:
-    // - Returns a function if there is a renderer *committed* to render;
-    // - Returns false if nothing should be rendered.
     return (...createRendererArgs) => {
       try {
         const render = runMiddleware(...createRendererArgs);
@@ -35,25 +39,24 @@ export function forRenderer(type, { strict = false } = {}, ...middleware) {
           }
 
           return <UserlandBoundary type={`render of ${type}`}>{render}</UserlandBoundary>;
-        } else {
-          return (...renderTimeArgs) => (
-            <UserlandBoundary type={`render of ${type}`}>
-              {() => {
-                try {
-                  const element = render(...renderTimeArgs);
-
-                  if (strict && !isValidElement(element)) {
-                    console.error(`botframework-webchat: ${type} should return React element only.`);
-                  }
-
-                  return element;
-                } catch (err) {
-                  return <ErrorBox error={err} type={type} />;
-                }
-              }}
-            </UserlandBoundary>
-          );
         }
+        return (...renderTimeArgs) => (
+          <UserlandBoundary type={`render of ${type}`}>
+            {() => {
+              try {
+                const element = render(...renderTimeArgs);
+
+                if (strict && !isValidElement(element)) {
+                  console.error(`botframework-webchat: ${type} should return React element only.`);
+                }
+
+                return element;
+              } catch (err) {
+                return <ErrorBox error={err} type={type} />;
+              }
+            }}
+          </UserlandBoundary>
+        );
       } catch (err) {
         return <ErrorBox error={err} type={type} />;
       }