+++ packages/component/src/Middleware/CardAction/createCoreMiddleware.js
@@ -1,3 +1,5 @@
+import { ie11 } from '../../Utils/detectBrowser';
+
 // This code is adopted from sanitize-html/naughtyScheme.
 // sanitize-html is a dependency of Web Chat but the naughtScheme function is neither exposed nor reusable.
 // https://github.com/apostrophecms/sanitize-html/blob/master/src/index.js#L526
@@ -45,7 +47,13 @@ export default function createDefaultCardActionMiddleware() {
         case 'playVideo':
         case 'showImage':
           if (ALLOWED_SCHEMES.includes(getScheme(value))) {
-            window.open(value, '_blank', 'noopener noreferrer');
+            if (ie11) {
+              const newWindow = window.open();
+              newWindow.opener = null;
+              newWindow.location = value;
+            } else {
+              window.open(value, '_blank', 'noopener noreferrer');
+            }
           } else {
             console.warn('botframework-webchat: Cannot open URL with disallowed schemes.', value);
           }
@@ -53,9 +61,12 @@ export default function createDefaultCardActionMiddleware() {
           break;
 
         case 'signin': {
-          // TODO: [P3] We should prime the URL into the OAuthCard directly, instead of calling getSessionId on-demand
-          //       This is to eliminate the delay between window.open() and location.href call
+          /**
+           * @todo TODO: [P3] We should prime the URL into the OAuthCard directly, instead of calling getSessionId on-demand
+           *       This is to eliminate the delay between window.open() and location.href call
+           */
 
+          // eslint-disable-next-line wrap-iife
           (async function () {
             const popup = window.open();
             const url = await getSignInUrl();