+++ packages/component/src/SendBox/SuggestedAction.js
@@ -17,13 +17,9 @@ const { useDirection, useDisabled, usePerformCardAction, useSuggestedActions } =
 
 const ROOT_STYLE = {
   '&.webchat__suggested-action': {
-    display: 'flex',
-    flexDirection: 'column',
-    whiteSpace: 'initial',
-
     '& .webchat__suggested-action__button': {
       display: 'flex',
-      overflow: 'hidden'
+      overflow: 'hidden' // Prevent image from leaking
     }
   }
 };
@@ -44,10 +40,12 @@ const connectSuggestedAction = (...selectors) =>
 const SuggestedAction = ({
   'aria-hidden': ariaHidden,
   buttonText,
+  className,
   displayText,
   image,
   imageAlt,
   text,
+  textClassName,
   type,
   value
 }) => {
@@ -81,7 +79,13 @@ const SuggestedAction = ({
   return (
     <div
       aria-hidden={ariaHidden}
-      className={classNames('webchat__suggested-action', rootClassName, suggestedActionStyleSet + '')}
+      className={classNames(
+        'webchat__suggested-action',
+        { 'webchat__suggested-action--rtl': direction === 'rtl' },
+        rootClassName,
+        suggestedActionStyleSet + '',
+        (className || '') + ''
+      )}
     >
       <AccessibleButton
         {...(accessKey ? { 'aria-keyshortcuts': localizeAccessKey(accessKey) } : {})}
@@ -101,7 +105,7 @@ const SuggestedAction = ({
             src={image}
           />
         )}
-        <nobr className="webchat__suggested-action__button-text">{buttonText}</nobr>
+        <span className={classNames('webchat__suggested-action__text', (textClassName || '') + '')}>{buttonText}</span>
       </AccessibleButton>
     </div>
   );
@@ -109,10 +113,12 @@ const SuggestedAction = ({
 
 SuggestedAction.defaultProps = {
   'aria-hidden': false,
+  className: '',
   displayText: '',
   image: '',
   imageAlt: undefined,
   text: '',
+  textClassName: '',
   type: '',
   value: undefined
 };
@@ -120,10 +126,12 @@ SuggestedAction.defaultProps = {
 SuggestedAction.propTypes = {
   'aria-hidden': PropTypes.bool,
   buttonText: PropTypes.string.isRequired,
+  className: PropTypes.string,
   displayText: PropTypes.string,
   image: PropTypes.string,
   imageAlt: PropTypes.string,
   text: PropTypes.string,
+  textClassName: PropTypes.string,
   type: PropTypes.string,
   value: PropTypes.any
 };