+++ packages/component/src/Styles/StyleSet/SendBoxTextBox.js
@@ -1,7 +1,9 @@
 export default function createSendBoxTextBoxStyle({
+  paddingRegular,
   primaryFont,
   sendBoxBackground,
   sendBoxDisabledTextColor,
+  sendBoxMaxHeight,
   sendBoxPlaceholderColor,
   sendBoxTextColor,
   subtle
@@ -10,16 +12,20 @@ export default function createSendBoxTextBoxStyle({
     '&.webchat__send-box-text-box': {
       alignItems: 'center',
       fontFamily: primaryFont,
+      padding: paddingRegular,
       position: 'relative',
 
       '& .webchat__send-box-text-box__input': {
-        backgroundColor: sendBoxBackground,
         border: 0,
         fontFamily: 'inherit',
         fontSize: 'inherit',
         height: '100%',
         outline: 0,
-        padding: 0,
+        padding: 0
+      },
+
+      '& .webchat__send-box-text-box__input, & .webchat__send-box-text-box__html-text-area': {
+        backgroundColor: sendBoxBackground,
 
         '&:not(:disabled):not([aria-disabled="true"])': {
           color: sendBoxTextColor
@@ -32,6 +38,10 @@ export default function createSendBoxTextBoxStyle({
         '&::placeholder': {
           color: sendBoxPlaceholderColor || subtle
         }
+      },
+
+      '& .webchat__send-box-text-box__input, & .webchat__send-box-text-box__text-area': {
+        maxHeight: sendBoxMaxHeight
       },
 
       '& .webchat__send-box-text-box__glass': {